requirements-lock.txt
# pip freeze output here (All dependencies included)

accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
bm3d==4.0.3
bm4d==4.2.5
certifi==2024.12.14
charset-normalizer==3.4.1
coloredlogs==15.0.1
compressai==1.2.6
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
diffusers==0.32.2
dill==0.3.8
einops==0.8.0
et_xmlfile==2.0.0
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
invisible-watermark==0.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lpips==0.1.4
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.20.1
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prettytable==3.12.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
pylibdmtx==0.1.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.1.10
tokenizers==0.21.0
torch==2.5.1
torch-geometric==2.6.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
